requirements.txt
numpy
opencv-python
matplotlib
scikit-image
Pillow
